// ts-tools/interfaces/Cw20Adapter.types.ts
/**
 * This file was automatically generated by @cosmwasm/ts-codegen@0.26.0.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run the @cosmwasm/ts-codegen generate command to regenerate this file.
 */

export interface InstantiateMsg {
  launchpad: string;
}
export type ExecuteMsg =
  | {
      receive: {
        amount: Uint128;
        msg: Binary;
        sender: string;
      };
    }
  | {
      redeem_and_transfer: {
        recipient?: string | null;
      };
    }
  | {
      update_metadata: {
        addr: Addr;
      };
    }
  | {
      register_r_g: {
        addr: Addr;
      };
    };
export type Uint128 = string;
export type Binary = string;
export type Addr = string;
export type QueryMsg = {
  new_denom_fee: {};
};
export type ArrayOfCoin = Coin[];
export interface Coin {
  amount: Uint128;
  denom: string;
  [k: string]: unknown;
}